% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_sim.R
\name{boot_sim}
\alias{boot_sim}
\title{Simulate from Survival Model using Seamon-Keogh Bootstrap Method}
\usage{
boot_sim(out, proc_inputs, quantiles, kwd, MMs)
}
\arguments{
\item{MMs}{all the Model Matrices passed as arguments in a list.}

\item{pars}{list of parameter settings to use}
}
\description{
Obtain samples from a specified Cox Marginal Structural Model
using the frugal parameterization.
}
