% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_survivl.R
\name{check_censoring}
\alias{check_censoring}
\title{Check for censoring among outcome formulas}
\usage{
check_censoring(formulas, pars, cns_kwd)
}
\arguments{
\item{formulas}{formulas only for outcome variables}

\item{pars}{list of parameters}

\item{cns_kwd}{word used to denote censoring}
}
\value{
A list containing (possibly) edited \code{formulas}, \code{pars}, and a
logical \code{censoring} indicating the presence of censoring.
}
\description{
Check for censoring among outcome formulas
}
