% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{clean_tms}
\alias{clean_tms}
\title{Cleans tms object in \code{process_inputs}}
\usage{
clean_tms(tms)
}
\arguments{
\item{tms}{a vector of tms (already flattened) that may have + ,_lk, or I(.) around them}
}
\description{
Cleans tms object in \code{process_inputs}
}
