% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helpers.R
\name{collect_events}
\alias{collect_events}
\title{Collect events}
\usage{
collect_events(dat, varY, t = 0L, trunc = FALSE)
}
\arguments{
\item{dat}{data frame of observations}

\item{varY}{character vector of response variable stems}

\item{t}{time index}

\item{trunc}{logical indicating whether variables have already been truncated}
}
\description{
Collect events
}
