% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{curr_inputs}
\alias{curr_inputs}
\title{Obtain reduced formulas and parameter vectors for earlier time-points}
\usage{
curr_inputs(formulas, pars, t, ordering, done, vars_t, kwd, start_at = 0)
}
\arguments{
\item{formulas}{list of formulas to use}

\item{pars}{list of parameter settings to use}

\item{t}{the current time-point}

\item{ordering}{an ordering for simulation of variables}

\item{done}{character vector of variables already simulated}

\item{vars_t}{stems for time-varying variables}

\item{kwd}{keyword for the copula}
}
\description{
Obtain reduced formulas and parameter vectors for earlier time-points
}
