% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_cdf_bootstrap.R
\name{empirical_cdf_bootstrap}
\alias{empirical_cdf_bootstrap}
\title{Simulate from Survival Model using Seamon-Keogh Bootstrap Method}
\usage{
empirical_cdf_bootstrap(mod_inputs, out, riskH)
}
\arguments{
\item{mod_inputs}{processed model inputs from \code{process_inputs}}

\item{out}{simulated data so far}

\item{riskH}{empirical cumulative hazard function used in the bootstrap}
}
\description{
Obtain samples from a specified Cox Marginal Structural Model
using the frugal parameterization.
}
