% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{get_surv_family}
\alias{get_surv_family}
\title{Obtain families including survival outcomes}
\usage{
get_surv_family(val)
}
\arguments{
\item{val}{integer corresponding to distributional family}
}
\description{
Obtain families including survival outcomes
}
\details{
The functions \code{gaussian_causl_fam()} etc. represent the functions that are
returned by \code{get_family()}.  This function also returns families defined
by the \code{survivl} package.
}
\seealso{
\link{family_vals}
}
