% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_funcs.R
\name{glm_ldens}
\alias{glm_ldens}
\title{Get log-density for rejection sampling}
\usage{
glm_ldens(Y, family, eta, phi, link, par2)
}
\arguments{
\item{Y}{observed value}

\item{family}{family for simulation}

\item{eta}{linear component}

\item{phi}{optional dispersion parameter}

\item{link}{link function}

\item{par2}{extra parameters}
}
\description{
Get log-density for rejection sampling
}
