% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_funcs.R
\name{glm_ldens}
\alias{glm_ldens}
\title{Get log-density for rejection sampling}
\usage{
glm_ldens(Y, family, eta, link, phi, other_pars)
}
\arguments{
\item{Y}{observed value}

\item{family}{family for simulation}

\item{eta}{linear component}

\item{link}{link function}

\item{phi}{optional dispersion parameter}

\item{other_pars}{list containing parameters specific to this \code{family}}
}
\value{
Returns a numeric vector of the same length as \code{eta}.
}
\description{
Get log-density for rejection sampling
}
\details{
Possible
families are Gaussian (=1), t (=2), Exponential (=3), beta (=4)
Bernoulli/categorical (=5) and log-normal (=6).

Link functions are listed in \code{causl::links_list}.
}
\section{Functions}{
\itemize{
\item \code{glm_ldens()}: Get log-density for rejection sampling

}}
