% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{is_surv_outcome}
\alias{is_surv_outcome}
\title{Check if family can model a time-to-event outcome}
\usage{
is_surv_outcome(x)
}
\arguments{
\item{x}{vector or list of families}
}
\description{
Check if family can model a time-to-event outcome
}
