% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{ll_surv}
\alias{ll_surv}
\title{Log-likelihood for survival data}
\usage{
ll_surv(dat, formulas, family, pars, link, control)
}
\arguments{
\item{dat}{data frame of class \code{survivl_dat}}

\item{formulas}{list of formulas to use}

\item{family}{list of families to use}

\item{pars}{list of parameter settings to use}

\item{link}{link functions for GLM-like models}

\item{control}{list of control parameters}
}
\description{
Evaluate the log-likelihood for a survival dataset.
}
\details{
The only valid argument for \code{control} is \code{cop}, which
changes the keyword for the copula from its default (\code{"cop"}).
}
