% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{manipulate_survival}
\alias{manipulate_survival}
\alias{surv_to_long}
\alias{surv_to_wide}
\title{Manipulate survival data}
\usage{
surv_to_long(dat, lag = 0)

surv_to_wide(dat, tv_covs, fix_covs)
}
\arguments{
\item{dat}{survival dataset}

\item{lag}{number of earlier time-points to include}

\item{tv_covs, fix_covs}{time-varying and fixed covariates}
}
\description{
Manipulate survival data
}
\details{
Functions to change the format of survival data as generated
by \code{coxSamp} or \code{coxSamp}.  The function \code{surv_to_long}
maps to one row for each time-point and individual, whereas \code{surv_to_wide}
has precisely one row for each individual.
}
\section{Functions}{
\itemize{
\item \code{surv_to_long()}: change to long format

\item \code{surv_to_wide()}: change to wide format

}}
