% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{mod_args}
\alias{mod_args}
\title{Modify conditional specified lagged formuals to correct formulas}
\usage{
mod_args(form, beta, t, modLHS = FALSE)
}
\arguments{
\item{form}{The conditional specified formulas with V_lk}

\item{beta}{the parameters specified}

\item{t}{the time point we are at currently}

\item{modLHS}{should we modify the left hand side variables in the formula?}
}
\description{
Modify conditional specified lagged formuals to correct formulas
}
