% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivl_model.R
\name{modify}
\alias{modify}
\title{Modify an object}
\usage{
modify(x, ...)
}
\arguments{
\item{x}{object to modify}

\item{...}{additional arguments passed to methods}
}
\description{
Generic function for modifying objects in-place.
}
