% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_survivl.R
\name{modify_inputs}
\alias{modify_inputs}
\title{Obtain only time-varying inputs}
\usage{
modify_inputs(proc_inputs)
}
\arguments{
\item{proc_inputs}{output of \code{process_inputs}}
}
\description{
Obtain only time-varying inputs
}
