% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivl_model.R
\name{print.survivl_model}
\alias{print.survivl_model}
\title{Display output from \code{survivl_model}}
\usage{
\method{print}{survivl_model}(x, ...)
}
\arguments{
\item{x}{an object of class \code{survivl_model}}

\item{...}{additional arguments (not used)}
}
\description{
Display output from \code{survivl_model}
}
