% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_survivl.R
\name{process_prespecified}
\alias{process_prespecified}
\title{Process variables already simulated}
\usage{
process_prespecified(dat, varC, vars_t, var_order, start = 0L)
}
\arguments{
\item{dat}{data frame of pre-generated variables}

\item{varC, vars_t}{character vector of static covariates and stems for dynamic variables}

\item{var_order}{integer giving order for dynamic variables}

\item{start}{time value to start simulation at}
}
\value{
A data frame of quantiles, and numbers representing the number of
time points already fully simulated (\code{upto}) and the number within the particular
group that have already been simulated (\code{vars_sim}).
}
\description{
Process variables already simulated
}
