% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{regex_extr}
\alias{regex_extr}
\title{Extract patterns from a string}
\usage{
regex_extr(pattern, x, sub_patt = "{")
}
\arguments{
\item{pattern}{regular expression to match}

\item{x}{string for pattern extraction; should contain a single capture group}

\item{sub_patt}{string to use for replacement}
}
\description{
Extract patterns from a string
}
\details{
This function is supposed to extract all captured elements from a
string, and return them as a character vector.  It actually returns the
whole match
}
