% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm_samp.R
\name{rmsm}
\alias{rmsm}
\title{Simulate from a \code{survivl_model} object}
\usage{
rmsm(n, surv_model, control = list())
}
\arguments{
\item{n}{number of samples}

\item{surv_model}{an object of class \code{survivl_model}}

\item{control}{list of control parameters}
}
\value{
An object of class \code{survivl_dat} containing the simulated data.
}
\description{
Obtain samples from a pre-specified \code{survivl_model} object.
}
