% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rmv_time_stamps}
\alias{rmv_time_stamps}
\alias{add_time_stamps}
\title{Add or remove time stamps to variable names}
\usage{
rmv_time_stamps(nms)

add_time_stamps(nms, t, start_at = 0)
}
\description{
Add or remove time stamps to variable names
}
\section{Functions}{
\itemize{
\item \code{add_time_stamps()}: replace lags with time stamps

}}
