% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bootstrap_survivl.R
\name{sim_bootstrap_survivl}
\alias{sim_bootstrap_survivl}
\title{Simulate survivl outcome using inversion method}
\usage{
sim_bootstrap_survivl(out, surv_model)
}
\arguments{
\item{out}{data frame for output}

\item{surv_model}{the object surv_model}
}
\description{
Simulate survivl outcome using inversion method
}
