% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_inversion_longitudinal.R
\name{sim_inversion_longitudinal}
\alias{sim_inversion_longitudinal}
\title{Simulate longitudinal non survival outcome using inversion method}
\usage{
sim_inversion_longitudinal(out, surv_model)
}
\arguments{
\item{out}{data frame for output}

\item{surv_model}{the object surv_model}
}
\description{
Simulate longitudinal non survival outcome using inversion method
}
