% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_rejection.R
\name{sim_rejection_survivl}
\alias{sim_rejection_survivl}
\title{Simulate survivl outcome using rejection method}
\usage{
sim_rejection_survivl(out, surv_model)
}
\arguments{
\item{out}{data frame for output}

\item{surv_model}{the object surv_model}
}
\description{
Simulate survivl outcome using rejection method
}
