% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_variable.R
\name{sim_variable}
\alias{sim_variable}
\title{Simulate a single variable using the inversion method. We note that
in Xi et al paper notation in the paper never has Y_0, treatment at time
t = 0 results in outcome at t = 1, Y_1. For organizational reasons, we
assume treatment at time t = 0 results in outcome at time t = 0 and so forth.}
\usage{
sim_variable(
  n,
  formulas,
  family,
  pars,
  link,
  dat,
  quantiles,
  type_event,
  num_time_varying
)
}
\arguments{
\item{n}{sample size}

\item{formulas}{list consisting of a formula for the output variables and a list of formulae for the pair-copula}

\item{family}{list containing family variable}

\item{pars}{list with two entries, first a list of parameters for response, and second a further list of parameters for pair-copula}

\item{link}{list of same form as \code{family}}

\item{dat}{data frame of current variables}

\item{quantiles}{data frame of quantiles}

\item{type_event}{character denoting if primary or competing}

\item{num_time_varying}{number of time varying covariates}
}
\value{
The data frame \code{dat} with an additional column given by the left-hand side of \code{formula[[1]]}.
}
\description{
Each entry \code{formulas}, \code{family}, \code{pars}, \code{link} is a list
with two entries, the first referring to the variable being simulated and the
second to the pair-copulas being used.
}
