% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\docType{data}
\name{surv_family_vals}
\alias{surv_family_vals}
\alias{surv_links_list}
\title{Numbers for parametric survival families}
\format{
\code{surv_family_vals} is a \code{data.frame} with 2 rows and 2 columns

\code{surv_links_list} is a list of length 2
}
\usage{
surv_family_vals

surv_links_list
}
\description{
Data frame containing
\itemize{
\item \code{val}: an integer
\item \code{family}: a vector giving the associated parametric family for that integer.
}
The integer \code{val} may be used in place of the name of the parametric family
when specifying the \code{family} object.  Note that the numbers are not currently
used, but this will be modified in a future version.
}
\section{Functions}{
\itemize{
\item \code{surv_links_list}: List of links for each family

}}
\keyword{datasets}
