% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{survivl_dat}
\alias{survivl_dat}
\title{Data from a survival model}
\description{
This class must have columns \code{id}, \code{T}, and \code{status}, for the
individual, their survival time and the form of their final event.  Static
covariates are given as single columns, while time-varying covariates are
suffixed by \code{_t}, where \code{t} is the relevant time-point.
}
