% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{survivl_fams}
\alias{survivl_fams}
\alias{weibull_surv_fam}
\title{Families for survival distributions}
\usage{
weibull_surv_fam(link)
}
\arguments{
\item{link}{link function to use}
}
\description{
For each family, the function returns a list containing a density,
simulation function, quantile function, and cumulative distribution
function.  These have the generic names \code{ddist}, \code{rdist}, \code{qdist} and
\code{pdist} respectively.
}
\section{Functions}{
\itemize{
\item \code{weibull_surv_fam()}: Weibull distribution

}}
