% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivl_model.R
\name{survivl_model}
\alias{survivl_model}
\title{Define a \code{survivl_model} object}
\usage{
survivl_model(
  formulas,
  family,
  pars,
  link,
  T = T,
  dat = NULL,
  qtls = NULL,
  method = "inversion",
  kwd = "cop",
  control = list()
)
}
\arguments{
\item{kwd}{word used for copula formula and parameters}
}
\description{
This defines a \code{survivl_model} object, that can be used either for simulation
or inference.
}
\details{
The components \code{formulas} and \code{family} must both be specified, and have
matching lengths.  If \code{pars} is specified, then the model can be used for
simulation and inference, if not then only for inference.  \code{link} is optional,
and if not specified default links will be used.
}
