% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{survSamp}
\alias{survSamp}
\title{Simulate from a survival model}
\usage{
survSamp(n, T, formulas, family, pars, link = NULL, control = list())
}
\arguments{
\item{n}{number of samples}

\item{T}{number of time-points}

\item{formulas}{formulae for model}

\item{family}{families of distributions}

\item{link}{list of link functions}

\item{control}{list of options for the algorithm}

\item{par}{list of parameter values}
}
\value{
A \code{data.frame} containing the variable on the left hand side
of each formula.
}
\description{
Simulate from a survival model
}
\details{
The formulas, family and (if specified) link arguments should be
lists with four entries; the first relates to any static baseline
covariates and the second to any (possibly) time-varying ones, the third
to any treatments, and the fourth to the outcome.  The four entries in
these lists should have the same length.

The function gamma() that controls the acceleration defaults to just a
linear combination of the covariates, but it can be adjusted to an arbitrary
function of them.

Link functions for the Gaussian, t and Gamma distributions can be the
identity, inverse or log functions.  Gaussian and t-distributions default to
the identity, and Gamma to the log link.  For the Bernoulli the logit and
probit links are available.

The function currently doesn't support varying the parameters at each
time-point.
}
\references{
Young, Hernan, Picciotto and Robins, 2008. Simulation from
Structural Survival Models under Complex Time-Varying Data Structures,
\emph{JSM Proceedings, Section on Statistics in Epidemiology: American Statistical Association}.
}
