% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{time_stamps}
\alias{time_stamps}
\alias{rmv_time_stamps}
\alias{add_time_stamps}
\title{Add or remove time stamps to variable names}
\usage{
rmv_time_stamps(nms)

add_time_stamps(nms, t, start_at = 0)
}
\arguments{
\item{nms}{names to add or remove time stamps from}

\item{t}{time point for adding}

\item{start_at}{first time point}
}
\description{
Add or remove time stamps to variable names
}
\section{Functions}{
\itemize{
\item \code{rmv_time_stamps()}: remove time stamps

\item \code{add_time_stamps()}: replace lags with time stamps

}}
