% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_cdf_bootstrap.R
\name{empirical_cdf_bootstrap}
\alias{empirical_cdf_bootstrap}
\title{Simulate from Survival Model using Seamon-Keogh Bootstrap Method}
\usage{
empirical_cdf_bootstrap(mod_inputs, out, riskH)
}
\arguments{
\item{out}{Simulated data so far.}

\item{pars}{list of parameter settings to use}
}
\description{
Obtain samples from a specified Cox Marginal Structural Model
using the frugal parameterization.
}
