% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_funcs.R
\name{glm_sim}
\alias{glm_sim}
\alias{glm_ldens}
\title{Simulate single variable}
\usage{
glm_sim(family, eta, phi, link, par2)

glm_ldens(Y, family, eta, link, phi, par2)
}
\arguments{
\item{family}{family for simulation}

\item{eta}{linear component}

\item{phi}{optional dispersion parameter}

\item{link}{link function}

\item{Y}{observed value}
}
\value{
Returns a numeric vector of the same length as \code{eta}.
}
\description{
Simulate single variable
}
\details{
Possible
families are Gaussian (=1), t (=2), Exponential (=3), beta (=4)
Bernoulli/categorical (=5) and log-normal (=6).

Link functions are listed in \code{causl:::linksList}.
}
\section{Functions}{
\itemize{
\item \code{glm_ldens()}: Get log-density for rejection sampling

}}
