% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{lag_data}
\alias{lag_data}
\title{Get lagged version of a dataset for a particular timepoint}
\usage{
lag_data(dat, t, var_nms, static = character(0))
}
\description{
Get lagged version of a dataset for a particular timepoint
}
