% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivl_model.R
\name{modify.survivl_model}
\alias{modify.survivl_model}
\title{Modify \code{survivl_model} object}
\usage{
modify.survivl_model(
  x,
  over = FALSE,
  formulas,
  family,
  pars,
  link,
  dat,
  method,
  kwd
)
}
\arguments{
\item{x}{an object of class \code{survivl_model}}

\item{over}{logical: should components be added/modified or entirely over-written?

This function can be used to modify}

\item{kwd}{word used for copula formula and parameters}
}
\description{
Change one or more components of a \code{survivl_model} object.
}
