% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{process_inputs}
\alias{process_inputs}
\title{Process formulas, families and parameters}
\usage{
process_inputs(n, formulas, pars, family, link, dat, T, method, control)
}
\arguments{
\item{n}{number of samples}

\item{formulas}{list of formulas to use}

\item{pars}{list of parameter settings to use}

\item{family}{list of families to use}

\item{link}{link functions for GLM-like models}

\item{dat}{optional data frame for plasmode simulation}

\item{T}{number of time points}

\item{method}{sampling method (defaults to \code{"inversion"})}

\item{control}{list of control parameters}
}
\description{
Process formulas, families and parameters
}
